Refactored RecursiveComparisonConfiguration to pull up reusable parts
